[SYCL] Fix int overflow for OffsetInBytes in accessors

+++ sycl/source/detail/accessor_impl.hpp
@@ -44,7 +44,7 @@ public:
   // TODO: Remove when ABI break is allowed.
   AccessorImplHost(id<3> Offset, range<3> AccessRange, range<3> MemoryRange,
                    access::mode AccessMode, void *SYCLMemObject, int Dims,
-                   int ElemSize, int OffsetInBytes = 0,
+                   int ElemSize, size_t OffsetInBytes = 0,
                    bool IsSubBuffer = false,
                    const property_list &PropertyList = {})
       : MAccData(Offset, AccessRange, MemoryRange), MAccessMode(AccessMode),
@@ -55,7 +55,7 @@ public:
 
   AccessorImplHost(id<3> Offset, range<3> AccessRange, range<3> MemoryRange,
                    access::mode AccessMode, void *SYCLMemObject, int Dims,
-                   int ElemSize, bool IsPlaceH, int OffsetInBytes = 0,
+                   int ElemSize, bool IsPlaceH, size_t OffsetInBytes = 0,
                    bool IsSubBuffer = false,
                    const property_list &PropertyList = {})
       : MAccData(Offset, AccessRange, MemoryRange), MAccessMode(AccessMode),
@@ -110,7 +110,7 @@ public:
 
   unsigned int MDims;
   unsigned int MElemSize;
-  unsigned int MOffsetInBytes;
+  size_t MOffsetInBytes;
   bool MIsSubBuffer;
 
   void *&MData = MAccData.MData;
